+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added <podcast:alternateEnclosure> support.
 - The ([podcast:license](https://github.com/vnali/craft-studio/discussions/9)) tag can be created via Asset fields now.
+- The ([podcast:trailer](https://github.com/vnali/craft-studio/discussions/15)) tag can be created via Asset fields now.
 
 ## 0.11.0 - 2023-07-03